Refactoring
* instantiate the NetworkServiceDiscovery with NsdManager instead of Context
