+++ nsd/src/main/java/de/lennartegb/nsd/NetworkServiceDiscovery.kt
@@ -1,5 +1,6 @@
 package de.lennartegb.nsd
 
+import de.lennartegb.nsd.model.NsdResult
 import kotlinx.coroutines.flow.Flow
 
 interface NetworkServiceDiscovery {
