+++ nsd/src/main/java/de/lennartegb/nsd/extensions/Context.kt
@@ -1,8 +1,11 @@
 package de.lennartegb.nsd.extensions
 
 import android.content.Context
+import android.net.nsd.NsdManager
 import de.lennartegb.nsd.NetworkServiceDiscovery
 import de.lennartegb.nsd.NetworkServiceDiscoveryImpl
 
-fun Context.getNetworkServiceDiscovery(): NetworkServiceDiscovery =
-	NetworkServiceDiscoveryImpl(this)
+fun Context.getNetworkServiceDiscovery(): NetworkServiceDiscovery {
+	val manager = getSystemService(Context.NSD_SERVICE) as NsdManager
+	return NetworkServiceDiscoveryImpl(manager)
+}
